feat: Style 컴포넌트 분리
